+++ package.json
@@ -1,21 +1,17 @@
 {
   "name": "tubular-fabric",
-  "version": "3.1.1",
-  "description": "Unosquare Tubular Fabric",
+  "version": "0.0.1",
+  "description": "Unosquare Tubular for Fabric",
   "main": "dist/index.js",
   "scripts": {
-    "build:docs": "parcel build ./srcdocs/index.html --no-source-maps --public-url /tubular-fabric/ -d ./docs",
     "build:sample": "parcel build ./sample/app/index.html --no-source-maps -d ./distsample",
-    "build": "tsc --p ./tsconfigs/tsconfig.tubular.json",
+    "build": "tsc --p ./tsconfig.json",
     "start": "parcel ./sample/app/index.html --open --port 8989 --no-source-maps -d ./distsample",
     "debug": "parcel ./sample/app/index.html --open --port 8989 -d ./distsample",
-    "start:docs": "parcel ./srcdocs/index.html --no-source-maps --port 8080 --open -d ./docs",
     "check": "npm run build && npm test",
-    "lint": "tsc --p ./tsconfigs/tsconfig.tubular.json --noEmit && eslint ./src/**/*.{js,ts,tsx} --quiet",
-    "lintfix": "tsc --p ./tsconfigs/tsconfig.tubular.json --noEmit && eslint ./src/**/*.{js,ts,tsx} --quiet --fix",
+    "lint": "tsc --p ./tsconfig.json --noEmit && eslint ./src/**/*.ts --quiet",
+    "lintfix": "tsc --p ./tsconfig.json --noEmit && eslint ./src/**/*.ts --quiet --fix",
     "test": "jest",
-    "postbuild:docs": "xcopy .\\srcdocs\\static .\\docs\\static\\",
-    "poststart:docs": "xcopy .\\srcdocs\\static .\\docs\\static\\",
     "preparepackage": "npm run build && npm run copypackage && cd dist && npm pack",
     "copypackage": "npx shx cp package.json dist"
   },
