+++ sample/src/ColumnsDefinition.ts
@@ -1,24 +1,46 @@
-import * as React from 'react';
-import { ColumnModel, ColumnSortDirection, ColumnDataType } from 'tubular-common';
+import { ColumnDataType } from 'tubular-common';
+import { ITbColumn } from '../../src/ITbColumn';
 
-export const columnsTR = [
-    new ColumnModel('OrderID', {
-        dataType: ColumnDataType.Numeric,
-        filterable: true,
-        isKey: true,
-        label: 'Id',
-        sortDirection: ColumnSortDirection.Ascending,
-        sortOrder: 1,
-        sortable: true,
-    }),
-    new ColumnModel('CustomerName', {
-        filterable: true,
-        searchable: true,
-        sortable: true,
-    }),
-    new ColumnModel('ShipperCity'),
-    new ColumnModel('Amount', {
-        dataType: ColumnDataType.Numeric,
-        sortable: true,
-    }),
+export const columns: ITbColumn[] = [
+    {
+        key: 'OrderID',
+        name: 'OrderID',
+        fieldName: 'OrderID',
+        minWidth: 100,
+        tb: {
+            isKey: true,
+            dataType: ColumnDataType.Numeric,
+            sortable: true,
+        },
+    },
+    {
+        key: 'CustomerName',
+        name: 'CustomerName',
+        fieldName: 'CustomerName',
+        minWidth: 100,
+        tb: {
+            dataType: ColumnDataType.String,
+            sortable: true,
+        },
+    },
+    {
+        key: 'ShipperCity',
+        name: 'ShipperCity',
+        fieldName: 'ShipperCity',
+        minWidth: 100,
+        tb: {
+            dataType: ColumnDataType.String,
+            sortable: true,
+        },
+    },
+    {
+        key: 'Amount',
+        name: 'Amount',
+        fieldName: 'Amount',
+        minWidth: 100,
+        tb: {
+            dataType: ColumnDataType.Numeric,
+            sortable: true,
+        },
+    },
 ];
